fix HintPath for nuget libraries
Restoring nuget packages from MonoDevelop restores them into the solution
directory instead of the project directory. Thus if the .csproj is used as
part of a different solution, the HintPaths are no longer resolvable.

Using $(SolutionDir) in the HintPath will make sure the path is correct for
the project itself and other solutions (like api-tests) that reference the
project.

This nuget issue is described here:
https://github.com/NuGet/Home/issues/738