fix timestamp format for API_REQUEST_DATE header on non-en_US locales
The API_REQUEST_DATE header has to be in the Unix timestamp format with
optional micro-seconds. If the micro-seconds are specified, a dot "." must be
used as the decimal point seperator.

The API server will reject timestamps that contain a culture dependent comma ","
as a separator. Double.ToString() uses by default a culture dependent saparator
though, e.g.:

    $ LANG=de_DE csharp -e 'var foo = 3.1d; foo.ToString();'
    "3,1"

So for the API server the serialization of the header needs to explicitely
specify the format in a culture independent way, e.g.:

    $ LANG=de_DE csharp -e 'var foo = 3.1d; foo.ToString("F3", System.Globalization.CultureInfo.InvariantCulture);'
    "3.100"

This way this API client library will correctly serialize the timestamps
regardless of what locale/culture the computer is using.